[WEB-3684] Parse out step name from step cvs
Whenever a step does not define a title, not the bitrise.yml provides one for the step
we simple use the source of the step to provide a simple name for the step instead of
using the full URL.

A simple name can be extracted by splitting the URL (both local and remote), and using
the last segment of it as a title.

To not lose information on the UI, the full URL can be provided to the user via the step's
tooltip, which makes it easier to know the actual source of the step by a glance.