[WEB-3684] Change StepItem's title to displayName
Since the title is a predefined HTML attribute, reusing it caused
an unexpected tooltip with wrong value to show up on the screen
beside the intended tooltip with the right value.

The usage of the React component had some wrongly cased properties
which did not raised any error, just silently broke a couple of things,
since the component got undefined as a value for the the wrongly cased
properties. It fixes the badge icon for example.

The tooltip was extended with the cvs to give users that information if needed.