[WEB-3412] fix(workflow-selector): Long names
In case the name is to long and two workflow differs at their ends,
the user was not able to distinguish them, due to hiding the information.

To make it more flexible the following changes were made:
- Allow word-wrapping to handler harder edge cases
- On desktop viewport (+580px) make the selector wider, to give more room

This required some fine-tuning of the current style:
- center the dot, whenever the name of the workflow wraps into new line
- make text behave more like when it is selected

Solution was to utilize `display: flex` on non selected item as well.