[Web-3684] Workflow editor ui bugs #plugin (#482)
* [WEB-3684] Add and extend step title tests

- Add "Fifth step" and "Eighth step" to `elements.js`
- Extend "wf2" with missing title steps, both github step (step 3) and local step (step 6) in `test_bitrise.yml`
- Fix "wf2" related tests in `workflows-steps-details.feature`
- Add github step missing title test under "Step title scenarios" in `workflows-steps-details.feature`
- Add local step overwritten title test under "Step title scenarios" in `workflows-steps-details.feature`
- Add local step missing test under "Step title scenarios" in `workflows-steps-details.feature`

* [WEB-3684] Parse out step name from step cvs

Whenever a step does not define a title, not the bitrise.yml provides one for the step
we simple use the source of the step to provide a simple name for the step instead of
using the full URL.

A simple name can be extracted by splitting the URL (both local and remote), and using
the last segment of it as a title.

To not lose information on the UI, the full URL can be provided to the user via the step's
tooltip, which makes it easier to know the actual source of the step by a glance.

* [WEB-3684] Change StepItem's title to displayName

Since the title is a predefined HTML attribute, reusing it caused
an unexpected tooltip with wrong value to show up on the screen
beside the intended tooltip with the right value.

The usage of the React component had some wrongly cased properties
which did not raised any error, just silently broke a couple of things,
since the component got undefined as a value for the the wrongly cased
properties. It fixes the badge icon for example.

The tooltip was extended with the cvs to give users that information if needed.

* (fix): Change required input border color to red

The border color was originally purple, however usually red
color indicates missing information to the user.

Fix required input border-color test as well.

* [WEB-3684] Modify tests

Due to changed requirements I had to modify the tests
to comply to them.

* [WEB-3684] test(Step): Add unit tests

- Add new tests to displayName
- Add new tests to displayCvs
- Add new tests to displayTooltip

* fix(_Step.js): Backport to ES5

Since we encourage TypeScript we make sure to feel the pain when we need to handle JavaScript files
to make the migration sooner to come, by limiting ourselves to ES5 intentionally.