test: Fix Gerkhin test warnings
**Problem**
_Cucumber (Gerkhin) Full Support_ VS Code plugin revealed a couple of issues in our .feature
files. Mostly wrong `When` - `And` and `Then` - `And` usage.

An `And` statement should be the same as it is chained to, basically is should be interchangeable
with the first statement's type in the chain:

Good usage
`When - And - And == When - When - When`
`Then - And - And == Then - Then - Then`

Bad usage
`When-Then-And != When-And-And`
`When-And-Then != When-And-And`

**Solution**
Install **Cucumber (Gerkhin) Full Support** (alexkrechik.cucumberautocomplete):
- Get intelligent autocompletion
- Get immediate validation and warnings